Adding the repository call and implementation of the service querying the dna info table

+++ src/main/java/com/meli/challenge/mutants/dto/DNAStats.java
@@ -6,7 +6,7 @@ import lombok.Data;
 @Data
 @Builder
 public class DNAStats {
-    private int countMutantDna;
-    private int countHumanDna;
+    private long countMutantDna;
+    private long countHumanDna;
     private double ratio;
 }
